Split nivo into separate component
