+++ packages/chart-react-nivo/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: "../react/.eslintrc.js",
+};
