+++ packages/chart-react-nivo/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@tony/cv-chart-react-nivo",
+  "license": "All rights reserved",
+  "private": true,
+  "version": "0.0.1",
+  "main": "dist/index.js",
+  "module": "src/index.ts",
+  "scripts": {
+    "start": "echo 'noop'",
+    "build": "echo 'noop'",
+    "test": "echo 'noop'",
+    "lint": "eslint . -c .eslintrc.js --ext .ts,.tsx",
+    "prettier": "prettier src webpack --write",
+    "format": "yarn run prettier",
+    "update": "yarn run ncu",
+    "ncu": "ncu",
+    "clean": "rm -rf dist/"
+  },
+  "devDependencies": {
+    "@types/node": "^14.14.21",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "cross-env": "^7.0.3",
+    "eslint-plugin-react": "^7.22.0",
+    "prettier": "^2.2.1",
+    "typescript": "^4.1.3"
+  },
+  "peerDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.13.0",
+    "@typescript-eslint/parser": "^4.13.0",
+    "eslint": "^7.18.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-promise": "^4.2.1"
+  },
+  "dependencies": {
+    "@datorama/akita": "^6.0.0",
+    "@nivo/bar": "^0.67.0",
+    "@nivo/core": "^0.67.0",
+    "@nivo/line": "^0.67.0",
+    "@nivo/pie": "^0.67.0",
+    "@tony/cv-lib": "*",
+    "fast-deep-equal": "^3.1.3",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "rxjs": "^6.6.3"
+  }
+}
