+++ packages/chart-react-nivo/src/hub.ts
@@ -1,0 +1,20 @@
+import {
+  activitiesQuery,
+  activityTypesQuery,
+  cvStore,
+  query,
+  orgsQuery,
+  orgTypesQuery,
+  languagesQuery,
+} from "@tony/cv-lib/hub";
+import { NivoChartQuery } from "./query";
+
+export const nivoChartQuery = new NivoChartQuery(
+  cvStore,
+  query,
+  activitiesQuery,
+  activityTypesQuery,
+  languagesQuery,
+  orgsQuery,
+  orgTypesQuery
+);
