+++ packages/chart-react-nivo/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../react/tsconfig.json",
+  "include": ["src/*"]
+}
