+++ packages/react/package.json
@@ -52,10 +52,6 @@
     "eslint-plugin-promise": "^4.2.1"
   },
   "dependencies": {
-    "@nivo/bar": "^0.67.0",
-    "@nivo/core": "^0.67.0",
-    "@nivo/line": "^0.67.0",
-    "@nivo/pie": "^0.67.0",
     "@tabler/icons": "^1.38.1",
     "fast-deep-equal": "^3.1.3",
     "html-webpack-plugin": "^4.5.1",
